+++ tiledesk-client/CHANGELOG.md
@@ -1,10 +1,12 @@
 # Client library for Tiledesk nodejs APIs client
 
 ### 0.7.15
-- 'project_id' instance property and constructor options property renamed in 'projectId'
+- 'project_id' instance property and constructor options property renamed in 'projectId' (camel case)
 - updateDepartment() renamed in updateRequestDepartment()
 - updateRequestDepartment: in options parameter added 'nobot' option
 - added updateProjectUser()
+- added documentation with jdoc
+- added getAllRequests()
 
 ### 0.7.14 - online
 - added getRequestById()